requirements.txt
torch
torch-geometric
scikit-learn
numpy
pandas
matplotlib
pyyaml